+++ src/main/java/com/github/sakaizawa/word2vec/Vector.java
@@ -14,27 +14,27 @@ public interface Vector {
      * ベクトルを返す
      * @return ベクトル
      */
-    public float[] getVector();
+    public double[] getVector();
 
     /**
      * 指定されたindexにnumberを加える
      * @param index
      * @param number 加える数
      */
-    public void add(int index, float number);
+    public void add(int index, double number);
 
     /**
      * ベクトルの大きさを返す
      * @return ベクトルの大きさ
      */
-    public float getSize();
+    public double getSize();
 
     /**
      * ベクトルの指定されたindexの値を返す
      * @param index
      * @return ベクトルの指定されたindexの値
      */
-    public float getElement(int index);
+    public double getElement(int index);
 
     /**
      * ベクトルの大きさを返す
